requirements.txt
google-generativeai==0.3.2
streamlit==1.28.0
python-dotenv==1.0.0 